avoid redeem dust
