+++ web/src/hooks/useWinningPositions.ts
@@ -1,10 +1,10 @@
 import { RouterAbi } from "@/abi/RouterAbi";
 import { generateWinningOutcomeIndexes } from "@/lib/conditional-tokens";
-import { Market } from "@/lib/market";
+import { Market, getRedeemedPrice } from "@/lib/market";
 import { config } from "@/wagmi";
 import { useQuery } from "@tanstack/react-query";
 import { readContract } from "@wagmi/core";
-import { Address } from "viem";
+import { Address, formatUnits } from "viem";
 import { Position, useMarketPositions } from "./useMarketPositions";
 
 export const useWinningPositions = (account: Address | undefined, market: Market, router: Address) => {
@@ -23,19 +23,25 @@ export const useWinningPositions = (account: Address | undefined, market: Market
 
       //  set balance=0 to positions that are not winning outcomes
       const winningPositionsWithZero = positions.map((position, i) => {
+        // Create a copy of the position to avoid modifying the original
+        const positionCopy = { ...position };
+
         if (market.type === "Generic") {
-          if (!winningOutcomes[i]) {
-            position.balance = 0n;
+          const redeemPrice = getRedeemedPrice(market, i);
+          const redeemAmount = redeemPrice * Number(formatUnits(position.balance, position.decimals));
+          if (redeemAmount < 0.01) {
+            // ignore this position if the amount to redeem is too small
+            positionCopy.balance = 0n;
           }
         } else {
           // futarchy market
           // if the winning outcome is the first one (YES), put to zero odd indexes (NO)
           // if the winning outcome is the second one (NO), put to zero even indexes (YES)
           if ((winningOutcomes[0] === true && i % 2 === 1) || (winningOutcomes[1] === true && i % 2 === 0)) {
-            position.balance = 0n;
+            positionCopy.balance = 0n;
           }
         }
-        return position;
+        return positionCopy;
       });
 
       // futarchy markets have 4 outcomes but only 2 of them can be winning positions (both YES or both NO)
